% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{description}
\alias{description}
\alias{description,Gene-method}
\title{Accessor function for the description of a Gene object.}
\usage{
description(x)

\S4method{description}{Gene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The description of the Gene

The updated Gene object.
}
\description{
Accessor function for the description of a Gene object.

Accessor function for the description of a Gene object.
}
\examples{
gene1 <- ProteinCodingGene(id=10447, symbol= 'FAM3C', name='FAM3
metabolism regulating signaling molecule C',description = 'encodes a secreted
protein with a GG domain', structure = GRanges("chr7",IRanges(121348878,
121396364), strand = '-'), proteinID = 'Q92520')
description(gene1)
}
