% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{final_mature_miRNAID}
\alias{final_mature_miRNAID}
\alias{final_mature_miRNAID,MiRNAGene-method}
\title{Accessor function for the final_mature_miRNAID of a micro-RNA Gene object.}
\usage{
final_mature_miRNAID(x)

\S4method{final_mature_miRNAID}{MiRNAGene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The final_mature_miRNAID slot value of the MiRNAGene object.

The updated Gene object.
}
\description{
Accessor function for the final_mature_miRNAID of a micro-RNA Gene object.

Accessor function for the final_mature_miRNAID of a micro-RNA object.
}
\examples{
gene3 <- MiRNAGene(id=406988, symbol= 'MIR205', name='MICRORNA 205',
description = 'non-coding RNAs that are involved in post-transcriptional
regulation of gene expression by affecting both the stability and translation
of mRNAs.', structure = GRanges("chr1", IRanges (209432133, 209432242),
strand = '+'), final_mature_miRNAID= 'hsa-miR-205')
final_mature_miRNAID(gene3)
}
