% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\title{Gene}
\description{
A general class for Genes with various attributes
}
\section{Slots}{

\describe{
\item{\code{id}}{Numeric vector specifying the unique identifier for the gene.}

\item{\code{symbol}}{Character vector specifying the symbol associated with the gene.}

\item{\code{name}}{Character vector specifying the name of the gene.}

\item{\code{description}}{Character vector providing additional description of gene.}

\item{\code{structure}}{GRanges object representing the genomic structure of gene.}
}}

