% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getseedseq}
\alias{getseedseq}
\alias{getseedseq,MiRNAGene-method}
\title{Get seed region for the MiRNA Gene.}
\usage{
getseedseq(x)

\S4method{getseedseq}{MiRNAGene}(x)
}
\arguments{
\item{x}{The MiRNAGene object.}
}
\value{
A seed region object.
}
\description{
This function retrieves the seed region for a MiRNA Gene.
}
\examples{
\dontrun{
gene3 <- MiRNAGene(id=406988, symbol= 'MIR205', name='MICRORNA 205',
description = 'non-coding RNAs that are involved in post-transcriptional
regulation of gene expression by affecting both the stability and translation
of mRNAs.', structure = GRanges("chr1", IRanges (209432133, 209432242),
strand = '+'), final_mature_miRNAID= 'hsa-miR-205')
data(hsSeqs)
getseedseq(gene3)
}
}
