% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lengthProduct,LncRNAGene-method}
\alias{lengthProduct,LncRNAGene-method}
\title{Calculate the length product for a long non-coding RNA gene.}
\usage{
\S4method{lengthProduct}{LncRNAGene}(x)
}
\arguments{
\item{x}{The LncRNAGene object.}
}
\value{
The length product of the gene structure.
}
\description{
This function calculates the length product for a LncRNAGene object,
which is the sum of the widths of its structure (GRanges) intervals.
}
\examples{
gene2 <- LncRNAGene(id=10984, symbol= 'KCNQ1OT1', name='antisense
transcript 1', description = 'interacts with chromatin and regulates
transcription of multiple target genes through epigenetic modifications',
structure = GRanges("chr11", IRanges(2608328, 2699994), strand = '-'),
lncRNAID='URS0000759CF4')
lengthProduct(gene2)
}
