% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lengthProduct,MiRNAGene-method}
\alias{lengthProduct,MiRNAGene-method}
\title{Calculate the length product for a micro-RNA gene.}
\usage{
\S4method{lengthProduct}{MiRNAGene}(x)
}
\arguments{
\item{x}{The MiRNAGene object.}
}
\value{
The length product of the micro-RNA gene.
}
\description{
This function calculates the length product for a MiRNAGene object,
which is the number of characters in the RNA sequence of the micro-RNA.
}
\examples{
\dontrun{
gene3 <- MiRNAGene(id=406988, symbol= 'MIR205',name='MICRORNA 205',
description = 'non-coding RNAs that are involved in post-transcriptional
regulation of gene expression by affecting both the stability and translation
of mRNAs.', structure = GRanges("chr1", IRanges (209432133, 209432242),
strand = '+'), final_mature_miRNAID= 'hsa-miR-205')
data(hsSeqs)
lengthProduct(gene3)
}
}
