% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lengthProduct,rRNAGene-method}
\alias{lengthProduct,rRNAGene-method}
\title{Calculate the length product for an rRNA gene.}
\usage{
\S4method{lengthProduct}{rRNAGene}(x)
}
\arguments{
\item{x}{The rRNAGene object.}
}
\value{
The length product of the rRNA gene.
}
\description{
This function calculates the length product for an rRNAGene object,
which is the sum of the widths of the gene structure.
}
\examples{
gene4 <- rRNAGene(id=100008588, symbol= 'RNA18SN5', name='RNA,
18S ribosomal N5',description = 'the 45S rDNA repeat unit encodes a 45S rRNA
precursor, transcribed by RNA polymerase I', structure = GRanges("chr21",
IRanges(8436876, 8438744), strand = "+"), rRNAID='URS0000726FAB')
lengthProduct(gene4)
}
