% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lengthProduct,SNRNAGene-method}
\alias{lengthProduct,SNRNAGene-method}
\title{Calculate the length product for an SNRNA gene.}
\usage{
\S4method{lengthProduct}{SNRNAGene}(x)
}
\arguments{
\item{x}{The SNRNAGene object.}
}
\value{
The length product of the SNRNA gene.
}
\description{
This function calculates the length product for an SNRNAGene object,
which is the sum of the widths of the gene structure.
}
\examples{
gene6 <- SNRNAGene(id=126871, symbol= 'RNU1-1', name='RNU1-1',
description = 'RNA, U1 small nuclear 1', structure = GRanges("chr1",
IRanges(16514122,16514285), strand = "-"), SNRNAID='URS000032B6B6')
lengthProduct(gene6)
}
