% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lengthProduct,tRNAGene-method}
\alias{lengthProduct,tRNAGene-method}
\title{Calculate the length product for a tRNA gene.}
\usage{
\S4method{lengthProduct}{tRNAGene}(x)
}
\arguments{
\item{x}{The tRNAGene object.}
}
\value{
The length product of the tRNA gene.
}
\description{
This function calculates the length product for a tRNAGene object,
which is the sum of the widths of the gene structure.
}
\examples{
gene5 <- tRNAGene(id=100189346, symbol= 'TRA-AGC1-1',name='tRNA-Ala
(anticodon AGC) 1-1', description = 'tRNA-Ala (anticodon AGC) 1-1',
structure = GRanges("chr6", IRanges(28795964, 28796035), strand = "-"),
tRNAID='URS000063E4FD_9606', anticodon='AGC')
lengthProduct(gene5)
}
