% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lengthProduct}
\alias{lengthProduct}
\alias{lengthProduct,ProteinCodingGene-method}
\title{Calculate the length product for a protein-coding gene.}
\usage{
lengthProduct(x)

\S4method{lengthProduct}{ProteinCodingGene}(x)
}
\arguments{
\item{x}{The ProteinCodingGene object.}
}
\value{
The length product of the protein sequences.
}
\description{
This function calculates the length product for a Coding Gene object,
which is the product of the lengths of all its protein sequences.
}
\examples{
gene1 <- ProteinCodingGene(id=10447, symbol= 'FAM3C', name='FAM3
metabolism regulating signaling molecule C',description = 'encodes a secreted
protein with a GG domain', structure = GRanges("chr7",IRanges(121348878,
121396364), strand = '-'), proteinID = 'Q92520')
lengthProduct(gene1)
}
