% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{LncRNAGene-class}
\alias{LncRNAGene-class}
\alias{LncRNAGene}
\title{LncRNAGene}
\description{
A subclass of gene representing long non-coding genes.
}
\section{Slots}{

\describe{
\item{\code{lncRNAID}}{The RNAcentral ID, character.}
}}

\examples{
gene2 <- LncRNAGene(id=10984, symbol= 'KCNQ1OT1', name='antisense transcript
1', description = 'interacts with chromatin and regulates transcription of
multiple target genes through epigenetic modifications', structure =
GRanges("chr11", IRanges(2608328, 2699994), strand = '-'),
lncRNAID='URS0000759CF4')
}
