% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{lncRNAID}
\alias{lncRNAID}
\alias{lncRNAID,LncRNAGene-method}
\title{Accessor function for the lncRNAID of a long non-coding RNA Gene object.}
\usage{
lncRNAID(x)

\S4method{lncRNAID}{LncRNAGene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The lncRNAID of the LncRNAGene

The updated Gene object.
}
\description{
Accessor function for the lncRNAID of a long non-coding RNA Gene object.

Accessor function for the lncRNAID of a long non-coding RNA object.
}
\examples{
gene2 <- LncRNAGene(id=10984, symbol= 'KCNQ1OT1', name='antisense
transcript 1', description = 'interacts with chromatin and regulates
transcription of multiple target genes through epigenetic modifications',
structure = GRanges("chr11", IRanges(2608328, 2699994), strand = '-'),
lncRNAID='URS0000759CF4')
lncRNAID(gene2)
}
