% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{LncRNAseq}
\alias{LncRNAseq}
\alias{LncRNAseq,LncRNAGene-method}
\title{Get the RNA sequence for the LncRNA Gene.}
\usage{
LncRNAseq(x)

\S4method{LncRNAseq}{LncRNAGene}(x)
}
\arguments{
\item{x}{The LncRNAGene object.}
}
\value{
An RNAString object representing the RNA sequence of the gene.
}
\description{
This function retrieves the RNA sequence for a LncRNAGene object.
}
\examples{
gene2 <- LncRNAGene(id=10984, symbol= 'KCNQ1OT1', name='antisense
transcript 1', description = 'interacts with chromatin and regulates
transcription of multiple target genes through epigenetic modifications',
structure = GRanges("chr11", IRanges(2608328, 2699994), strand = '-'),
lncRNAID='URS0000759CF4')
LncRNAseq(gene2)
}
