% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{matureMiRNAseq}
\alias{matureMiRNAseq}
\alias{matureMiRNAseq,MiRNAGene-method}
\title{Get the mature RNA sequence for the MiRNA Gene class.}
\usage{
matureMiRNAseq(x)

\S4method{matureMiRNAseq}{MiRNAGene}(x)
}
\arguments{
\item{x}{The MiRNAGene object.}
}
\value{
An RNAString object representing the RNA sequence.
}
\description{
This function retrieves the mature RNA sequence for a MiRNA Gene.
}
\examples{
\dontrun{
gene3 <- MiRNAGene(id=406988, symbol= 'MIR205', name='MICRORNA 205',
description = 'non-coding RNAs that are involved in post-transcriptional
regulation of gene expression by affecting both the stability and translation
of mRNAs.', structure = GRanges("chr1", IRanges (209432133, 209432242),
strand = '+'), final_mature_miRNAID= 'hsa-miR-205')
data(hsSeqs)
matureMiRNAseq(gene3)
}
}
