% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{MiRNAGene-class}
\alias{MiRNAGene-class}
\alias{MiRNAGene}
\title{MiRNAGene}
\description{
A subclass of gene representing microRNA genes.
}
\section{Slots}{

\describe{
\item{\code{final_mature_miRNAID}}{The miRBase Sanger ID, character.}
}}

\examples{
gene3 <- MiRNAGene(id=406988, symbol= 'MIR205', name='MICRORNA 205',
description = 'non-coding RNAs that are involved in post-transcriptional
regulation of gene expression by affecting both the stability and translation
of mRNAs.', structure = GRanges("chr1", IRanges (209432133, 209432242),
strand = '+'), final_mature_miRNAID= 'hsa-miR-205')
}
