% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{mRNAseq}
\alias{mRNAseq}
\alias{mRNAseq,ProteinCodingGene-method}
\title{Get the mature mRNA sequence for the Coding Gene}
\usage{
mRNAseq(x)

\S4method{mRNAseq}{ProteinCodingGene}(x)
}
\arguments{
\item{x}{The ProteinCodingGene object.}
}
\value{
An RNAStringSetList object representing the mature
mRNA sequences of the gene.
}
\description{
This function retrieves the mature mRNA sequence for
a Protein Coding Gene object.
}
\examples{
gene1 <- ProteinCodingGene(id=10447, symbol= 'FAM3C', name='FAM3
metabolism regulating signaling molecule C',description = 'encodes a secreted
protein with a GG domain', structure = GRanges("chr7",IRanges(121348878,
121396364), strand = '-'), proteinID = 'Q92520')
mRNAseq(gene1)
}
