% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{ProteinCodingGene-class}
\alias{ProteinCodingGene-class}
\alias{ProteinCodingGene}
\title{ProteinCodingGene}
\description{
A subclass of gene representing protein-coding genes.
}
\section{Slots}{

\describe{
\item{\code{proteinID}}{The Uniprot ID, character.}
}}

\examples{
gene1 <- ProteinCodingGene(id=10447, symbol= 'FAM3C', name='FAM3 metabolism
regulating signaling molecule C',description = 'encodes a secreted protein
with a GG domain', structure = GRanges("chr7",IRanges(121348878, 121396364),
strand = '-'), proteinID = 'Q92520')
}
