% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{proteinID}
\alias{proteinID}
\alias{proteinID,ProteinCodingGene-method}
\title{Accessor function for the proteinID of a Coding Gene object.}
\usage{
proteinID(x)

\S4method{proteinID}{ProteinCodingGene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The proteinID of the Coding Gene

The updated Gene object.
}
\description{
Accessor function for the proteinID of a Coding Gene object.

Accessor function for the proteinID of a Coding Gene object.
}
\examples{
gene1 <- ProteinCodingGene(id=10447, symbol= 'FAM3C', name='FAM3
metabolism regulating signaling molecule C',description = 'encodes a secreted
protein with a GG domain', structure = GRanges("chr7",IRanges(121348878,
121396364), strand = '-'), proteinID = 'Q92520')
proteinID(gene1)
}
