% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{proteinseq}
\alias{proteinseq}
\alias{proteinseq,ProteinCodingGene-method}
\title{Get the protein sequence for the Protein Coding Gene class.}
\usage{
proteinseq(x)

\S4method{proteinseq}{ProteinCodingGene}(x)
}
\arguments{
\item{x}{The Protein Coding Gene object.}
}
\value{
A list of protein sequences corresponding to the exonic
regions of the gene.
}
\description{
This function retrieves the protein sequence for Protein Coding Gene object.
}
\examples{
gene1 <- ProteinCodingGene(id=10447, symbol= 'FAM3C', name='FAM3 metabolism
regulating signaling molecule C',description = 'encodes a secreted protein
with a GG domain', structure = GRanges("chr7",IRanges(121348878, 121396364),
strand = '-'), proteinID = 'Q92520')
proteinseq(gene1)
}
