% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{rRNAGene-class}
\alias{rRNAGene-class}
\alias{rRNAGene}
\title{rRNAGene}
\description{
A subclass of gene representing ribosomal RNA genes.
}
\section{Slots}{

\describe{
\item{\code{rRNAID}}{The RNAcentral ID, character.}
}}

\examples{
gene4 <- rRNAGene(id=100008588, symbol= 'RNA18SN5', name='RNA, 18S ribosomal
N5',description = 'the 45S rDNA repeat unit encodes a 45S rRNA precursor,
transcribed by RNA polymerase I', structure = GRanges("chr21",
IRanges(8436876, 8438744), strand = "+"), rRNAID='URS0000726FAB')
}
