% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{rRNAID}
\alias{rRNAID}
\alias{rRNAID,rRNAGene-method}
\title{Accessor function for the rRNAID of an ribosomial RNA Gene object.}
\usage{
rRNAID(x)

\S4method{rRNAID}{rRNAGene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The rRNAID slot value of the rRNAGene object.

The updated Gene object.
}
\description{
Accessor function for the rRNAID of an ribosomial RNA Gene object.

Accessor function for the rRNAID of a ribosomial RNA object.
}
\examples{
gene4 <- rRNAGene(id=100008588, symbol= 'RNA18SN5', name='RNA,
18S ribosomal N5',description = 'the 45S rDNA repeat unit encodes a 45S rRNA
precursor, transcribed by RNA polymerase I', structure = GRanges("chr21",
IRanges(8436876, 8438744), strand = "+"), rRNAID='URS0000726FAB')
rRNAID(gene4)
}
