% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{rRNAseq}
\alias{rRNAseq}
\alias{rRNAseq,rRNAGene-method}
\title{Get the RNA sequence for the rRNA Gene}
\usage{
rRNAseq(x)

\S4method{rRNAseq}{rRNAGene}(x)
}
\arguments{
\item{x}{The rRNAGene object.}
}
\value{
An RNAString object representing the RNA sequence of the gene.
}
\description{
This function retrieves the RNA sequence for an rRNAGene object.
}
\examples{
gene4 <- rRNAGene(id=100008588, symbol= 'RNA18SN5', name='RNA,
18S ribosomal N5',description = 'the 45S rDNA repeat unit encodes a 45S rRNA
precursor, transcribed by RNA polymerase I', structure = GRanges("chr21",
IRanges(8436876, 8438744), strand = "+"), rRNAID='URS0000726FAB')
rRNAseq(gene4)
}
