% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{SNRNAGene-class}
\alias{SNRNAGene-class}
\alias{SNRNAGene}
\title{SNRNAGene}
\description{
A subclass of gene representing small-nuclear RNA genes.
}
\section{Slots}{

\describe{
\item{\code{SNRNAID}}{The RNAcentral ID, character.}
}}

\examples{
gene6 <- SNRNAGene(id=126871, symbol= 'RNU1-1', name='RNU1-1', description =
'RNA, U1 small nuclear 1', structure = GRanges("chr1", IRanges(16514122,
16514285), strand = "-"), SNRNAID='URS000032B6B6')
}
