% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{SNRNAID}
\alias{SNRNAID}
\alias{SNRNAID,SNRNAGene-method}
\title{Accessor function for SNRNAID slot in small-nuclear RNA Gene objects.}
\usage{
SNRNAID(x)

\S4method{SNRNAID}{SNRNAGene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The value of the SNRNAID slot.

The updated Gene object.
}
\description{
Accessor function for SNRNAID slot in small-nuclear RNA Gene objects.

Accessor function for the SNRNAID of a small-nuclear RNA object.
}
\examples{
gene6 <- SNRNAGene(id=126871, symbol= 'RNU1-1', name='RNU1-1',
description = 'RNA, U1 small nuclear 1', structure = GRanges("chr1",
IRanges(16514122,16514285), strand = "-"), SNRNAID='URS000032B6B6')
SNRNAID(gene6)
}
