% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SNRNAseq}
\alias{SNRNAseq}
\alias{SNRNAseq,SNRNAGene-method}
\title{Get the RNA sequence for the SNRNA Gene.}
\usage{
SNRNAseq(x)

\S4method{SNRNAseq}{SNRNAGene}(x)
}
\arguments{
\item{x}{The SNRNAGene object.}
}
\value{
An RNAString object representing the RNA sequence of the gene.
}
\description{
This function retrieves the RNA sequence for an SNRNAGene object.
}
\examples{
gene6 <- SNRNAGene(id=126871, symbol= 'RNU1-1', name='RNU1-1',
description = 'RNA, U1 small nuclear 1', structure = GRanges("chr1",
IRanges(16514122, 16514285), strand = "-"), SNRNAID='URS000032B6B6')
SNRNAseq(gene6)
}
