% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{tRNAGene-class}
\alias{tRNAGene-class}
\alias{tRNAGene}
\title{tRNAGene}
\arguments{
\item{object}{An instance of the tRNAGene class to be validated.}
}
\value{
TRUE if the object is valid, an error message otherwise.
}
\description{
A subclass of gene representing transfer RNA genes.
}
\details{
This validity function checks if the anticodon slot contains only
ribo-nucleotides and if it is 3 nucleotides long.
If the anticodon fails any of these checks, an appropriate error message is
returned; otherwise, TRUE is returned.
}
\section{Slots}{

\describe{
\item{\code{tRNAID}}{The RNAcentral ID, character.}

\item{\code{tRNAID}}{A character vector specifying the tRNA identifier.}

\item{\code{anticodon}}{A character vector specifying the anticodon sequence of tRNA.}
}}

\examples{
gene5 <- tRNAGene(id=100189346, symbol= 'TRA-AGC1-1', name='tRNA-Ala
(anticodon AGC) 1-1', description = 'tRNA-Ala (anticodon AGC) 1-1',
structure = GRanges("chr6", IRanges(28795964, 28796035), strand = "-"),
tRNAID='URS000063E4FD_9606', anticodon='AGC')
}
