% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\name{tRNAID}
\alias{tRNAID}
\alias{tRNAID,tRNAGene-method}
\title{Accessor function for the tRNAID of a transfer RNA Gene object.}
\usage{
tRNAID(x)

\S4method{tRNAID}{tRNAGene}(x)
}
\arguments{
\item{x}{A Gene object.}
}
\value{
The tRNAID slot value of the transfer RNA Gene object.

The updated Gene object.
}
\description{
Accessor function for the tRNAID of a transfer RNA Gene object.

Accessor function for the tRNAID of a transfer RNA object.
}
\examples{
gene5 <- tRNAGene(id=100189346, symbol= 'TRA-AGC1-1',name='tRNA-Ala
(anticodon AGC) 1-1', description = 'tRNA-Ala (anticodon AGC) 1-1',
structure = GRanges("chr6", IRanges(28795964, 28796035), strand = "-"),
tRNAID='URS000063E4FD_9606', anticodon='AGC')
tRNAID(gene5)
}
