% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{tRNAseq}
\alias{tRNAseq}
\alias{tRNAseq,tRNAGene-method}
\title{Get the RNA sequence for the tRNA Gene.}
\usage{
tRNAseq(x)

\S4method{tRNAseq}{tRNAGene}(x)
}
\arguments{
\item{x}{The tRNAGene object.}
}
\value{
An RNAString object representing the RNA sequence of the gene.
}
\description{
This function retrieves the RNA sequence for a tRNAGene object.
}
\examples{
gene5 <- tRNAGene(id=100189346, symbol= 'TRA-AGC1-1',name='tRNA-Ala
(anticodon AGC) 1-1', description = 'tRNA-Ala (anticodon AGC) 1-1',
structure = GRanges("chr6", IRanges(28795964, 28796035), strand = "-"),
tRNAID='URS000063E4FD_9606', anticodon='AGC')
tRNAseq(gene5)
}
